Resolve "Initialize weights on device side"

+++ HugeCTR/src/network.cu
@@ -19,7 +19,8 @@
 
 namespace HugeCTR {
 
-void conv_weight_gpu(int grid, int block, __half* dst, float* src, int elems, cudaStream_t stream) {
+void conv_weight_gpu(size_t grid, size_t block, __half* dst, float* src, int elems,
+                     cudaStream_t stream) {
   convert_array<<<grid, block, 0, stream>>>(dst, src, elems);
 }
